Refactor CSS
This PR removes the Vite demo project CSS and ports over the CSS from
the old book site. It also adds Sass and moves the components into
individual folders to more easily group components with their CSS
files and tests.

Fixes #15
