+++ package.json
@@ -38,6 +38,7 @@
     "jest": "26.6.3",
     "npm-run-all": "4.1.5",
     "prettier": "2.3.2",
+    "sass": "^1.37.5",
     "stylelint": "13.13.1",
     "stylelint-config-prettier": "8.0.2",
     "stylelint-config-spaceninja": "7.0.0",
