--- src/assets/main.css
@@ -1,404 +1,0 @@
-:root {
-  --custom-font-family: -apple-system, blinkmacsystemfont, segoe ui, roboto,
-    oxygen, ubuntu, cantarell, fira sans, droid sans, helvetica neue, sans-serif;
-  --custom-bg-color: #101010;
-  --custom-panel-color: #222222;
-  --custom-box-shadow: 0 2px 8px 0 rgba(0, 0, 0, 0.8);
-  --custom-color: #ffffff;
-  --custom-color-brand: #24b47e;
-  --custom-color-secondary: #666666;
-  --custom-color-alert: rgba(255, 255, 0, 0.25);
-  --custom-color-error: rgba(255, 0, 0, 0.25);
-  --custom-border: 1px solid #333333;
-  --custom-border-radius: 5px;
-  --custom-spacing: 5px;
-  --avatar-size: 10em;
-  background-color: var(--custom-bg-color);
-  font-family: var(--custom-font-family);
-  margin: 0;
-  padding: 0;
-}
-
-* {
-  box-sizing: border-box;
-  color: var(--custom-color);
-  font-family: var(--custom-font-family);
-}
-
-/* stylelint-disable-next-line selector-max-id */
-#__next,
-html,
-body {
-  height: 100vh;
-  overflow-x: hidden;
-  width: 100vw;
-}
-
-/* Grid */
-
-.container {
-  margin-left: auto;
-  margin-right: auto;
-  padding: 50px 0 100px;
-  width: 90%;
-}
-.row {
-  position: relative;
-  width: 100%;
-}
-.row [class^='col'] {
-  float: left;
-  margin: 0.5rem 2%;
-  min-height: 0.125rem;
-}
-.col-1,
-.col-2,
-.col-3,
-.col-4,
-.col-5,
-.col-6,
-.col-7,
-.col-8,
-.col-9,
-.col-10,
-.col-11,
-.col-12 {
-  width: 96%;
-}
-.col-1-sm {
-  width: 4.33%;
-}
-.col-2-sm {
-  width: 12.66%;
-}
-.col-3-sm {
-  width: 21%;
-}
-.col-4-sm {
-  width: 29.33%;
-}
-.col-5-sm {
-  width: 37.66%;
-}
-.col-6-sm {
-  width: 46%;
-}
-.col-7-sm {
-  width: 54.33%;
-}
-.col-8-sm {
-  width: 62.66%;
-}
-.col-9-sm {
-  width: 71%;
-}
-.col-10-sm {
-  width: 79.33%;
-}
-.col-11-sm {
-  width: 87.66%;
-}
-.col-12-sm {
-  width: 96%;
-}
-.row::after {
-  clear: both;
-  content: '';
-  display: table;
-}
-.hidden-sm {
-  display: none;
-}
-
-@media only screen and (min-width: 33.75em) {
-  /* 540px */
-  .container {
-    width: 80%;
-  }
-}
-
-@media only screen and (min-width: 45em) {
-  /* 720px */
-  .col-1 {
-    width: 4.33%;
-  }
-  .col-2 {
-    width: 12.66%;
-  }
-  .col-3 {
-    width: 21%;
-  }
-  .col-4 {
-    width: 29.33%;
-  }
-  .col-5 {
-    width: 37.66%;
-  }
-  .col-6 {
-    width: 46%;
-  }
-  .col-7 {
-    width: 54.33%;
-  }
-  .col-8 {
-    width: 62.66%;
-  }
-  .col-9 {
-    width: 71%;
-  }
-  .col-10 {
-    width: 79.33%;
-  }
-  .col-11 {
-    width: 87.66%;
-  }
-  .col-12 {
-    width: 96%;
-  }
-  .hidden-sm {
-    display: block;
-  }
-}
-
-@media only screen and (min-width: 60em) {
-  /* 960px */
-  .container {
-    max-width: 60rem;
-    width: 75%;
-  }
-}
-
-/* Forms */
-
-label {
-  color: var(--custom-color-secondary);
-  display: block;
-  font-size: 0.8rem;
-  margin: 5px 0;
-  text-transform: uppercase;
-}
-
-input {
-  background-color: var(--custom-bg-color);
-  border: var(--custom-border);
-  border-radius: 5px;
-  color: var(--custom-color);
-  font-size: 0.9rem;
-  padding: 8px;
-  width: 100%;
-}
-
-input[disabled] {
-  color: var(--custom-color-secondary);
-}
-
-textarea {
-  background-color: var(--custom-bg-color);
-  border: var(--custom-border);
-  border-radius: 5px;
-  color: var(--custom-color);
-  font-size: 0.9rem;
-  padding: 8px;
-  width: 100%;
-}
-
-/* Utils */
-
-.block {
-  display: block;
-  width: 100%;
-}
-.inline-block {
-  display: inline-block;
-  width: 100%;
-}
-.flex {
-  display: flex;
-}
-.flex.column {
-  flex-direction: column;
-}
-.flex.row {
-  flex-direction: row;
-}
-.flex.flex-1 {
-  flex: 1 1 0;
-}
-.flex-end {
-  justify-content: flex-end;
-}
-.flex-center {
-  justify-content: center;
-}
-.items-center {
-  align-items: center;
-}
-.text-sm {
-  font-size: 0.8rem;
-  font-weight: 300;
-}
-.text-right {
-  text-align: right;
-}
-.font-light {
-  font-weight: 300;
-}
-.opacity-half {
-  opacity: 50%;
-}
-
-/* Button */
-
-button,
-.button {
-  background-color: var(--custom-bg-color);
-  border: var(--custom-border);
-  border-radius: var(--custom-border-radius);
-  color: var(--custom-color);
-  cursor: pointer;
-  display: inline-block;
-  font-size: 0.9rem;
-  padding: 0.5rem 1rem;
-  text-align: center;
-  text-transform: uppercase;
-}
-
-/* stylelint-disable-next-line selector-no-qualifying-type */
-button.primary,
-.button.primary {
-  background-color: var(--custom-color-brand);
-  border: 1px solid var(--custom-color-brand);
-}
-
-/* Widgets */
-
-.card {
-  border: var(--custom-border);
-  border-radius: var(--custom-border-radius);
-  display: block;
-  margin: 1em 0;
-  padding: var(--custom-spacing);
-  width: 100%;
-}
-
-.alert {
-  background-color: var(--custom-color-alert);
-  border-radius: var(--custom-border-radius);
-  display: block;
-  margin: 1em 0;
-  padding: var(--custom-spacing);
-  width: 100%;
-}
-
-.alert--error {
-  background-color: var(--custom-color-error);
-}
-
-.avatar {
-  border-radius: var(--custom-border-radius);
-  max-width: 100%;
-  overflow: hidden;
-}
-.avatar.image {
-  object-fit: cover;
-}
-.avatar.no-image {
-  background-color: #333333;
-  border: 1px solid rgb(200, 200, 200);
-  border-radius: 5px;
-}
-.avatar--large {
-  height: var(--avatar-size);
-  width: var(--avatar-size);
-}
-
-.footer {
-  background-color: var(--custom-bg-color);
-  border-top: var(--custom-border);
-  bottom: 0;
-  display: flex;
-  flex-flow: row;
-  left: 0;
-  max-width: 100%;
-  position: absolute;
-  right: 0;
-}
-.footer div {
-  align-items: center;
-  display: flex;
-  padding: var(--custom-spacing);
-  width: 100%;
-}
-.footer div > img,
-.footer div > svg {
-  height: 20px;
-  margin-left: 10px;
-}
-.footer > div:first-child {
-  display: none;
-}
-.footer > div:nth-child(2) {
-  justify-content: left;
-}
-
-@media only screen and (min-width: 60em) {
-  /* 960px */
-  .footer > div:first-child {
-    display: flex;
-  }
-  .footer > div:nth-child(2) {
-    justify-content: center;
-  }
-}
-
-@keyframes spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.mainHeader {
-  font-size: 1.3rem;
-  margin-bottom: 20px;
-  width: 100%;
-}
-
-.avatarPlaceholder {
-  align-items: center;
-  background-color: rgba(255, 255, 255, 0.2);
-  border: var(--custom-border);
-  border-radius: var(--custom-border-radius);
-  display: flex;
-  height: 35px;
-  justify-content: center;
-  width: 35px;
-}
-
-.form-widget {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.form-widget > .button {
-  align-items: center;
-  background-color: #444444;
-  border: none;
-  display: flex;
-  justify-content: center;
-  text-transform: none !important;
-  transition: all 0.2s ease;
-}
-
-.form-widget .button:hover {
-  background-color: #2a2a2a;
-}
-
-.form-widget .button > .loader {
-  animation: spin 1s linear infinite;
-  filter: invert(1);
-  width: 17px;
-}
