+++ src/components/AppAlert/AppAlert.scss
@@ -1,0 +1,15 @@
+.alert {
+  --color-alert-warning: #ffffcc;
+  --color-alert-danger: #ffcccc;
+  border: 1px solid currentColor;
+  margin: 1em 0;
+  padding: 0.5em 1em;
+}
+
+.alert--warning {
+  background: var(--color-alert-warning);
+}
+
+.alert--error {
+  background: var(--color-alert-danger);
+}
