+++ src/components/AppButton/AppButton.scss
@@ -1,0 +1,55 @@
+$color_blue: #007bff;
+$color_dark_blue: #0069d9;
+$color_secondary: #6c757d;
+$color_secondary_dark: #5a6268;
+
+.button {
+  border: none;
+  border-radius: 0.25em;
+  cursor: pointer;
+  display: inline-block;
+  font-size: 16px;
+  padding: 0.375em 0.75em;
+
+  + .button {
+    margin-left: 1em;
+  }
+}
+
+.button__icon {
+  color: transparent;
+  text-shadow: 0 0 var(--color-button-text);
+  transform: scale(-1, 1);
+}
+
+.button--sm {
+  font-size: 14px;
+}
+
+.button--icon-only {
+  border-radius: 50%;
+  font-size: 12px;
+  height: 2.75em;
+  padding: 0;
+  width: 2.75em;
+}
+
+.button--primary {
+  background: $color_blue;
+  color: var(--color-button-text);
+
+  &:hover,
+  &:focus {
+    background: $color_dark_blue;
+  }
+}
+
+.button--secondary {
+  background: $color_secondary;
+  color: var(--color-button-text);
+
+  &:hover,
+  &:focus {
+    background: $color_secondary_dark;
+  }
+}
