+++ src/components/AppButton/AppButton.vue
@@ -7,3 +7,7 @@
 <script>
 export default {};
 </script>
+
+<style lang="scss">
+@import './AppButton';
+</style>
