--- src/components/AppLoading.vue
@@ -1,51 +1,0 @@
-<template>
-  <div class="loader">Loading...</div>
-</template>
-
-<script>
-export default {};
-</script>
-
-<style>
-.loader,
-.loader::before,
-.loader::after {
-  animation: load7 1.8s infinite ease-in-out;
-  animation-fill-mode: both;
-  border-radius: 50%;
-  height: 2.5em;
-  width: 2.5em;
-}
-.loader {
-  animation-delay: -0.16s;
-  color: currentColor;
-  font-size: 10px;
-  margin: 80px auto;
-  position: relative;
-  text-indent: -9999em;
-  transform: translateZ(0);
-}
-.loader::before,
-.loader::after {
-  content: '';
-  position: absolute;
-  top: 0;
-}
-.loader::before {
-  animation-delay: -0.32s;
-  left: -3.5em;
-}
-.loader::after {
-  left: 3.5em;
-}
-@keyframes load7 {
-  0%,
-  80%,
-  100% {
-    box-shadow: 0 2.5em 0 -1.3em;
-  }
-  40% {
-    box-shadow: 0 2.5em 0 0;
-  }
-}
-</style>
