+++ src/components/AvatarImage/AvatarImage.vue
