+++ src/components/BookCard/BookCard.scss
@@ -1,0 +1,168 @@
+$color_purchased: rgba(0, 128, 0, 0.33);
+$color_prioritized: rgba(255, 0, 0, 0.33);
+$color_warning: #ffffcc;
+$color_danger: #ffcccc;
+$color_shadow1: rgba(0, 0, 0, 0.12);
+$color_shadow2: rgba(0, 0, 0, 0.24);
+
+.book {
+  background: var(--color-book-bg);
+  border-radius: 3px;
+  box-shadow: 0 1px 3px $color_shadow1, 0 1px 2px $color_shadow2;
+  display: flex;
+  margin: 0;
+  padding: 1em 1em 0.5em;
+  position: relative;
+
+  > * + * {
+    margin-left: 1em;
+  }
+
+  &::before {
+    background: var(--color-rating-bg);
+    border-bottom-left-radius: 4px;
+    border-top-right-radius: 4px;
+    content: '';
+    display: block;
+    height: 5.25rem;
+    position: absolute;
+    right: 0;
+    top: 0;
+    width: 4.5rem;
+  }
+
+  &::after {
+    border: 10px solid transparent;
+    border-top-left-radius: 4px;
+    content: '';
+    height: 0;
+    left: 0;
+    position: absolute;
+    top: 0;
+    width: 0;
+  }
+}
+
+.book--purchased {
+  &::after {
+    border-left-color: $color_purchased;
+    border-top-color: $color_purchased;
+  }
+}
+
+.book--prioritized {
+  &::after {
+    border-left-color: $color_prioritized;
+    border-top-color: $color_prioritized;
+  }
+
+  &.book--purchased {
+    &::after {
+      border-left-color: $color_purchased;
+    }
+  }
+}
+
+.book__thumbnail {
+  flex: none;
+  max-width: 25%;
+  width: 50px;
+
+  @media screen and (min-width: 640px) {
+    width: 100px;
+  }
+}
+
+.book__thumbnail-media {
+  display: block;
+  height: auto;
+  width: 100%;
+}
+
+.book__info {
+  flex: 1;
+
+  > :first-child {
+    margin-top: 0;
+  }
+
+  > :last-child {
+    margin-bottom: 0;
+  }
+}
+
+.book__headline {
+  color: var(--color-headline-text);
+  font-size: 1.25em;
+  font-weight: 600;
+  line-height: 1.2;
+  margin-bottom: 0.33em;
+  margin-right: 4rem;
+
+  @media screen and (min-width: 640px) {
+    font-size: 1.5em;
+  }
+}
+
+.book__series {
+  font-weight: 300;
+  opacity: 0.5;
+}
+
+.book__author {
+  line-height: 1.2;
+  margin-right: 4rem;
+  opacity: 0.7;
+}
+
+.book__rating {
+  font-size: 2em;
+  font-weight: 200;
+  line-height: 1;
+  opacity: 0.5;
+  position: absolute;
+  right: 0;
+  text-align: center;
+  top: 1rem;
+  width: 4.5rem;
+}
+
+.book__length {
+  line-height: 1;
+  opacity: 0.5;
+  position: absolute;
+  right: 0;
+  text-align: center;
+  top: 3.5rem;
+  width: 4.5rem;
+}
+
+.book__snippet {
+  overflow-wrap: break-word; // 2
+  word-wrap: break-word; // 1
+}
+
+.book__recommendation {
+  background: var(--color-recommendation-bg);
+  margin: 0 -0.5em;
+  opacity: 0.7;
+  padding: 0.4em 0.75em;
+}
+
+.book__note {
+  font-style: italic;
+  overflow-wrap: break-word; // 2
+  word-wrap: break-word; // 1
+}
+
+.book__actions {
+  display: flex;
+  flex: none;
+  flex-direction: column;
+  padding-top: 5rem;
+
+  .button + .button {
+    margin-left: 0;
+    margin-top: 0.5em;
+  }
+}
