+++ src/components/BookDev/BookDev.vue
@@ -9,5 +9,5 @@ import {
   deleteAllBooks,
   addInitBooks,
   addSampleBooks,
-} from '../composables/useBookDev';
+} from '../../composables/useBookDev';
 </script>
