+++ src/components/BookList/BookList.scss
@@ -1,0 +1,16 @@
+.book-list {
+  display: grid;
+  grid-gap: 1em;
+  grid-template-columns: 1fr;
+  list-style: none;
+  margin: 1em 0;
+  padding: 0;
+
+  @media screen and (min-width: 640px) {
+    grid-template-columns: 1fr 1fr;
+  }
+
+  @media screen and (min-width: 1024px) {
+    grid-template-columns: repeat(auto-fill, minmax(375px, 1fr));
+  }
+}
