+++ src/components/FormInput/FormInput.scss
@@ -1,0 +1,80 @@
+.form {
+  display: grid;
+  grid-gap: 1em;
+  grid-template-columns: 1fr;
+  margin: 1em auto;
+
+  @media only screen and (min-width: 640px) {
+    grid-template-columns: 1fr 1fr;
+  }
+}
+
+.form__group--doublewide {
+  @media only screen and (min-width: 640px) {
+    grid-column-end: span 2;
+  }
+}
+
+.form__control {
+  background: var(--color-input-bg);
+  border: 1px solid var(--color-input-border);
+  border-radius: 0.25em;
+  color: currentColor;
+  display: block;
+  font-size: 16px;
+  padding: 0.375em 0.75em;
+  width: 100%;
+}
+
+select {
+  font-size: 16px;
+}
+
+.form__group--check {
+  align-items: center;
+  display: flex;
+}
+
+.form__checkbox-input {
+  flex: none;
+}
+
+.form__checkbox-label {
+  flex: 1;
+  margin-left: 0.375em;
+}
+
+.form__input-group {
+  display: flex;
+
+  .form__input-group-prepend,
+  .form__input-group-append {
+    background: var(--color-input-append-bg);
+    border: 1px solid var(--color-input-border);
+    flex: none;
+    padding: 0.375em 0.75em;
+  }
+
+  .form__input-group-prepend {
+    border-right: 0;
+  }
+
+  .form__input-group-append {
+    border-left: 0;
+  }
+
+  .form__control {
+    border-radius: 0;
+    flex: 1;
+  }
+
+  :first-child {
+    border-bottom-left-radius: 0.25em;
+    border-top-left-radius: 0.25em;
+  }
+
+  :last-child {
+    border-bottom-right-radius: 0.25em;
+    border-top-right-radius: 0.25em;
+  }
+}
