+++ src/components/RouterLink/RouterLink.vue
@@ -16,7 +16,7 @@ defineProps({
 </script>
 
 <script>
-import { routes } from '../routes';
+import { routes } from '../../routes';
 
 export default {
   methods: {
