+++ src/pages/Home.vue
@@ -6,5 +6,5 @@
 
 <script setup>
 import MainLayout from '../layouts/Main.vue';
-import BookList from '../components/BookList.vue';
+import BookList from '../components/BookList/BookList.vue';
 </script>
