+++ src/pages/Profile.vue
@@ -6,5 +6,5 @@
 
 <script setup>
 import MainLayout from '../layouts/Main.vue';
-import ProfileEdit from '../components/ProfileEdit.vue';
+import ProfileEdit from '../components/ProfileEdit/ProfileEdit.vue';
 </script>
